Minor tweaks and updated readme files.  Ensured all scripts ran in GroovyConsole and worked correctly

+++ src/test/groovy/gCudaScripts/cScalerMultiply.cu
@@ -2,7 +2,7 @@ extern "C"
 __global__ void scalerMultiply (int s,int vSize,float *vector) {
   int i = blockIdx.x * blockDim.x + threadIdx.x;
   if ( i < vSize){
-    vector[i] = vector[i] * s;
+    vector[i] = (float) (vector[i] * s);
   }
 } // scalerMultiply
